Fix for MISSING_RESULT error
Oops: I was using self._cache.get(key) to lookup the key, which returns
None if the key isn't found - the client then interprets None as an
error, since it isn't a string.

Changed this to use MemcacheGetResult, which has both the fetched value
and a flag indicating whether it was found. This allows the client to
distinguish between those cases. (Another way might be to use an
optional; I'll test that next.)